restructure orchestrator bin to use control thread
